Add files via upload
A simple database project that stores events planned on some day that suppors the following operations: 
Add - Adds events given by date. Format is the following: Add YYYY/MM/DD event.
Del - deletes all the events if only date is given, if event is given too, then deletes the event on a date. Format is the following: Del YYYY/MM/DD event. event is optional.
Print - prints all the data found in a database. Format is straight forward: Print
Find - finds all events on a particular date as the first input.  Format is the following: Find YYYY/MM/DD.
Database ignores wrong inputs if keywords written not as provided in the instruction, and tells if date format is wrong.